fixed contract formatting

+++ contracts/src/SupplyChain.sol
@@ -14,16 +14,16 @@ contract SupplyChain {
     struct Shipment {
         address sender;
         address receiver;
-        uint pickupTime;
-        uint deliveryTime;
-        uint distance;
-        uint price;
+        uint256 pickupTime;
+        uint256 deliveryTime;
+        uint256 distance;
+        uint256 price;
         ShipmentStatus status;
         bool isPaid;
     }
 
     // errors
-    error PaymentDoesNotMatchPrice(uint amount, uint price, string note);
+    error PaymentDoesNotMatchPrice(uint256 amount, uint256 price, string note);
     error InvalidReceiver();
     error ShipmentNotInIntendedStatus(string note);
     error ShipmentIsAlreadypaid();
@@ -32,16 +32,16 @@ contract SupplyChain {
     mapping(address => Shipment[]) public shipments;
 
     // number of shipments
-    uint public shipmentCount;
+    uint256 public shipmentCount;
 
     // displayed shipment
     struct TestShipment {
         address sender;
         address receiver;
-        uint pickupTime;
-        uint deliveryTime;
-        uint distance;
-        uint price;
+        uint256 pickupTime;
+        uint256 deliveryTime;
+        uint256 distance;
+        uint256 price;
         ShipmentStatus status;
         bool isPaid;
     }
@@ -51,61 +51,28 @@ contract SupplyChain {
     // events
 
     event ShipmentCreated(
-        address indexed sender,
-        address indexed receiver,
-        uint pickupTime,
-        uint distance,
-        uint price
+        address indexed sender, address indexed receiver, uint256 pickupTime, uint256 distance, uint256 price
     );
 
-    event ShipmentInTransit(
-        address indexed sender,
-        address indexed receiver,
-        uint pickupTime
-    );
+    event ShipmentInTransit(address indexed sender, address indexed receiver, uint256 pickupTime);
 
-    event ShipmentDelivered(
-        address indexed sender,
-        address indexed receiver,
-        uint deliveryTime
-    );
-    event ShipmentPaid(
-        address indexed sender,
-        address indexed receiver,
-        uint amount
-    );
+    event ShipmentDelivered(address indexed sender, address indexed receiver, uint256 deliveryTime);
+    event ShipmentPaid(address indexed sender, address indexed receiver, uint256 amount);
 
     constructor() {
         shipmentCount = 0;
     }
 
-    function createShipment(
-        address _receiver,
-        uint _pickupTime,
-        uint _distance,
-        uint _price
-    ) public payable {
+    function createShipment(address _receiver, uint256 _pickupTime, uint256 _distance, uint256 _price) public payable {
         // checks the actual value being sent (such as 10 ETH) matches the price the user sets
         if (msg.value != _price) {
-            revert PaymentDoesNotMatchPrice(
-                msg.value,
-                _price,
-                "Payment does not match the price user set"
-            );
+            revert PaymentDoesNotMatchPrice(msg.value, _price, "Payment does not match the price user set");
         }
 
         // creates a temporary shipment structure, which then is pushed into mapping(address => Shipment) shipments
         // when creating the shipment: delivery time is 0, shipment status is pending (it was just intialised), and is has not been paid for
-        Shipment memory shipment = Shipment(
-            msg.sender,
-            _receiver,
-            _pickupTime,
-            0,
-            _distance,
-            _price,
-            ShipmentStatus.PENDING,
-            false
-        );
+        Shipment memory shipment =
+            Shipment(msg.sender, _receiver, _pickupTime, 0, _distance, _price, ShipmentStatus.PENDING, false);
 
         // pushes the created shipment struct into main mapping linked to the users address
         shipments[msg.sender].push(shipment);
@@ -114,35 +81,16 @@ contract SupplyChain {
 
         // test shipment too
         testShipments.push(
-            TestShipment(
-                msg.sender,
-                _receiver,
-                _pickupTime,
-                0,
-                _distance,
-                _price,
-                ShipmentStatus.PENDING,
-                false
-            )
+            TestShipment(msg.sender, _receiver, _pickupTime, 0, _distance, _price, ShipmentStatus.PENDING, false)
         );
 
         // emits an event for the shipment created
-        emit ShipmentCreated(
-            msg.sender,
-            _receiver,
-            _pickupTime,
-            _distance,
-            _price
-        );
+        emit ShipmentCreated(msg.sender, _receiver, _pickupTime, _distance, _price);
     }
 
     // once the shipment is created, it is held within the contract, and the shipping proccess will be started using this function
     // index refers to the ID. The user adds the index when wanting to start the shipment they created
-    function startShipment(
-        address _sender,
-        address _receiver,
-        uint _index
-    ) public {
+    function startShipment(address _sender, address _receiver, uint256 _index) public {
         // fetches shipment we created, puts it in storage state as we will change its state on the blockchain
         // we change its status
         // explaination of code: we go into the shipments array, and get the shipment associated with the sender and index(ID) we want.
@@ -153,9 +101,7 @@ contract SupplyChain {
             revert InvalidReceiver();
         }
         if (shipment.status != ShipmentStatus.PENDING) {
-            revert ShipmentNotInIntendedStatus(
-                "Shipment is not in Pending state"
-            );
+            revert ShipmentNotInIntendedStatus("Shipment is not in Pending state");
         }
 
         // changes shipment status to in-transit phase
@@ -167,11 +113,7 @@ contract SupplyChain {
     }
 
     // once in-transit has finished, it is delivered, and we must complete it
-    function completeShipment(
-        address _sender,
-        address _receiver,
-        uint _index
-    ) public {
+    function completeShipment(address _sender, address _receiver, uint256 _index) public {
         Shipment storage shipment = shipments[_sender][_index];
         TestShipment storage testShipment = testShipments[_index];
 
@@ -179,9 +121,7 @@ contract SupplyChain {
             revert InvalidReceiver();
         }
         if (shipment.status != ShipmentStatus.IN_TRANSIT) {
-            revert ShipmentNotInIntendedStatus(
-                "Shipment is not in In-Transit state"
-            );
+            revert ShipmentNotInIntendedStatus("Shipment is not in In-Transit state");
         }
         if (shipment.isPaid) {
             revert ShipmentIsAlreadypaid();
@@ -195,7 +135,7 @@ contract SupplyChain {
         testShipment.deliveryTime = block.timestamp;
 
         // once the delivery process is complete, we complete payment to the sender (such as a manufacturer)
-        uint amount = shipment.price;
+        uint256 amount = shipment.price;
         payable(shipment.sender).transfer(amount);
         shipment.isPaid = true;
         testShipment.isPaid = true;
@@ -205,13 +145,10 @@ contract SupplyChain {
     }
 
     // get shipment for display
-    function getShipment(
-        address _sender,
-        uint _index
-    )
+    function getShipment(address _sender, uint256 _index)
         public
         view
-        returns (address, address, uint, uint, uint, uint, ShipmentStatus, bool)
+        returns (address, address, uint256, uint256, uint256, uint256, ShipmentStatus, bool)
     {
         // we are simply fetching the shipment for display, no state changes
         Shipment memory shipment = shipments[_sender][_index];
@@ -228,7 +165,7 @@ contract SupplyChain {
     }
 
     // gets the number of shipments made by a user
-    function getShipmentCount(address _sender) public view returns (uint) {
+    function getShipmentCount(address _sender) public view returns (uint256) {
         return shipments[_sender].length;
     }
