+++ contracts/test/SupplyChain.t.sol
@@ -16,27 +16,22 @@ contract SupplyChainTest is Test {
     }
 
     function testCreateShipment() public {
-        uint pickupTime = 12345;
-        uint distance = 10;
-        uint price = 1 ether;
+        uint256 pickupTime = 12345;
+        uint256 distance = 10;
+        uint256 price = 1 ether;
 
         vm.startPrank(sender);
-        supplyChain.createShipment{value: 1 ether}(
-            receiver,
-            pickupTime,
-            distance,
-            price
-        );
+        supplyChain.createShipment{value: 1 ether}(receiver, pickupTime, distance, price);
 
         vm.stopPrank();
 
         (
             address _sender,
             address _receiver,
-            uint _pickupTime,
-            uint _deliveryTime,
-            uint _distance,
-            uint _price,
+            uint256 _pickupTime,
+            uint256 _deliveryTime,
+            uint256 _distance,
+            uint256 _price,
             SupplyChain.ShipmentStatus _status,
             bool _isPaid
         ) = supplyChain.getShipment(sender, 0);
@@ -47,60 +42,44 @@ contract SupplyChainTest is Test {
         assertEq(_deliveryTime, _deliveryTime);
         assertEq(_distance, distance);
         assertEq(_price, price);
-        assertEq(uint(_status), uint(SupplyChain.ShipmentStatus.PENDING));
+        assertEq(uint256(_status), uint256(SupplyChain.ShipmentStatus.PENDING));
         assertEq(_isPaid, false);
     }
 
     function testAttemptToCreateShipmentButInsufficientEthSent() public {
-        uint pickupTime = 12345;
-        uint distance = 10;
-        uint price = 1 ether;
+        uint256 pickupTime = 12345;
+        uint256 distance = 10;
+        uint256 price = 1 ether;
 
         vm.startPrank(sender);
         vm.expectRevert();
-        supplyChain.createShipment{value: 0.5 ether}(
-            receiver,
-            pickupTime,
-            distance,
-            price
-        );
+        supplyChain.createShipment{value: 0.5 ether}(receiver, pickupTime, distance, price);
         vm.stopPrank();
     }
 
     function testCreateAndStartShipment() public {
-        uint pickupTime = 12345;
-        uint distance = 10;
-        uint price = 1 ether;
+        uint256 pickupTime = 12345;
+        uint256 distance = 10;
+        uint256 price = 1 ether;
 
         vm.startPrank(sender);
-        supplyChain.createShipment{value: 1 ether}(
-            receiver,
-            pickupTime,
-            distance,
-            price
-        );
+        supplyChain.createShipment{value: 1 ether}(receiver, pickupTime, distance, price);
 
         supplyChain.startShipment(sender, receiver, 0);
         vm.stopPrank();
 
-        (, , , , , , SupplyChain.ShipmentStatus _status, ) = supplyChain
-            .getShipment(sender, 0);
+        (,,,,,, SupplyChain.ShipmentStatus _status,) = supplyChain.getShipment(sender, 0);
 
-        assertEq(uint(_status), uint(SupplyChain.ShipmentStatus.IN_TRANSIT));
+        assertEq(uint256(_status), uint256(SupplyChain.ShipmentStatus.IN_TRANSIT));
     }
 
     function testCreateAndAttemptToStartShipmentButWrongReceiver() public {
-        uint pickupTime = 12345;
-        uint distance = 10;
-        uint price = 1 ether;
+        uint256 pickupTime = 12345;
+        uint256 distance = 10;
+        uint256 price = 1 ether;
 
         vm.startPrank(sender);
-        supplyChain.createShipment{value: price}(
-            receiver,
-            pickupTime,
-            distance,
-            price
-        );
+        supplyChain.createShipment{value: price}(receiver, pickupTime, distance, price);
 
         vm.expectRevert();
         supplyChain.startShipment(sender, sender, 0);
@@ -108,17 +87,12 @@ contract SupplyChainTest is Test {
     }
 
     function testCreateAndAttemptToStartShipmentButWrongStage() public {
-        uint pickupTime = 12345;
-        uint distance = 10;
-        uint price = 1 ether;
+        uint256 pickupTime = 12345;
+        uint256 distance = 10;
+        uint256 price = 1 ether;
 
         vm.startPrank(sender);
-        supplyChain.createShipment{value: price}(
-            receiver,
-            pickupTime,
-            distance,
-            price
-        );
+        supplyChain.createShipment{value: price}(receiver, pickupTime, distance, price);
 
         supplyChain.startShipment(sender, receiver, 0);
         supplyChain.completeShipment(sender, receiver, 0);
@@ -128,51 +102,30 @@ contract SupplyChainTest is Test {
     }
 
     function testCreateAndStartAndCompleteShipment() public {
-        uint pickupTime = 12345;
-        uint distance = 10;
-        uint price = 1 ether;
+        uint256 pickupTime = 12345;
+        uint256 distance = 10;
+        uint256 price = 1 ether;
 
         vm.startPrank(sender);
-        supplyChain.createShipment{value: 1 ether}(
-            receiver,
-            pickupTime,
-            distance,
-            price
-        );
+        supplyChain.createShipment{value: 1 ether}(receiver, pickupTime, distance, price);
 
         supplyChain.startShipment(sender, receiver, 0);
         supplyChain.completeShipment(sender, receiver, 0);
         vm.stopPrank();
 
-        (
-            ,
-            ,
-            ,
-            ,
-            ,
-            ,
-            SupplyChain.ShipmentStatus _status,
-            bool _isPaid
-        ) = supplyChain.getShipment(sender, 0);
+        (,,,,,, SupplyChain.ShipmentStatus _status, bool _isPaid) = supplyChain.getShipment(sender, 0);
 
         assertEq(_isPaid, true);
-        assertEq(uint(_status), uint(SupplyChain.ShipmentStatus.DELIVERED));
+        assertEq(uint256(_status), uint256(SupplyChain.ShipmentStatus.DELIVERED));
     }
 
-    function testCreateAndStartAndAttemptCompleteShipmentButWrongReceiver()
-        public
-    {
-        uint pickupTime = 12345;
-        uint distance = 10;
-        uint price = 1 ether;
+    function testCreateAndStartAndAttemptCompleteShipmentButWrongReceiver() public {
+        uint256 pickupTime = 12345;
+        uint256 distance = 10;
+        uint256 price = 1 ether;
 
         vm.startPrank(sender);
-        supplyChain.createShipment{value: price}(
-            receiver,
-            pickupTime,
-            distance,
-            price
-        );
+        supplyChain.createShipment{value: price}(receiver, pickupTime, distance, price);
 
         vm.expectRevert();
         supplyChain.startShipment(sender, sender, 0);
@@ -182,26 +135,16 @@ contract SupplyChainTest is Test {
     }
 
     function testGetAllTransaction() public {
-        uint pickupTime = 12345;
-        uint distance = 10;
-        uint price = 1 ether;
+        uint256 pickupTime = 12345;
+        uint256 distance = 10;
+        uint256 price = 1 ether;
 
         vm.startPrank(sender);
-        supplyChain.createShipment{value: 1 ether}(
-            receiver,
-            pickupTime,
-            distance,
-            price
-        );
+        supplyChain.createShipment{value: 1 ether}(receiver, pickupTime, distance, price);
         vm.stopPrank();
 
         vm.startPrank(sender);
-        supplyChain.createShipment{value: 10 ether}(
-            receiver,
-            pickupTime,
-            distance,
-            10 ether
-        );
+        supplyChain.createShipment{value: 10 ether}(receiver, pickupTime, distance, 10 ether);
         vm.stopPrank();
 
         assertEq(2, supplyChain.getAllTransactions().length);
